pyproject.toml
[tool.poetry]
name = "jira-issue-creator"
version = "0.1.0"
description = ""
authors = ["José Victor Ferreira <jose.ferreira@consulting-for.edenred.com>"]

[tool.poetry.dependencies]
python = "^3.8"
holidays = "^0.9.11"
jira = "^2.0"
fire = "^0.2.1"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
neovim = "^0.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
